DATAGRAPH-389: Refactoring- Moved some methods into superclass
